+++ bot/helper/mirror_leech_utils/download_utils/rclone_download.py
@@ -59,7 +59,11 @@ class RcloneLeech:
     
     async def cancel_download(self):
         self.__is_cancelled = True
-        self.process.kill() 
+        if self.process is not None:
+            try:
+                self.process.kill()
+            except:
+                pass
         await self.__listener.onDownloadError('Download cancelled!')
